More tests and refactoring
Extract code shared with the `Transport.Jobs.GtfsToDB` job.
