+++ apps/transport/lib/registry/gtfs.ex
@@ -6,6 +6,8 @@ defmodule Transport.Registry.GTFS do
   alias Transport.Registry.Model.Stop
   alias Transport.Registry.Model.StopIdentifier
 
+  alias Transport.GTFS.Utils
+
   require Logger
 
   @behaviour Transport.Registry.Extractor
@@ -23,7 +25,7 @@ defmodule Transport.Registry.GTFS do
 
         stops =
           content
-          |> to_stream_of_maps()
+          |> Utils.to_stream_of_maps()
           |> Stream.flat_map(&handle_stop/1)
           |> Enum.to_list()
 
@@ -31,30 +33,9 @@ defmodule Transport.Registry.GTFS do
     end
   end
 
-  @doc """
-  Transform the stream outputed by Unzip to a stream of maps, each map
-  corresponding to a row from the CSV.
-  """
-  def to_stream_of_maps(file_stream) do
-    file_stream
-    # transform the stream to a stream of binaries
-    |> Stream.map(fn c -> IO.iodata_to_binary(c) end)
-    # stream line by line
-    |> NimbleCSV.RFC4180.to_line_stream()
-    |> NimbleCSV.RFC4180.parse_stream(skip_headers: false)
-    # transform the stream to a stream of maps %{column_name1: value1, ...}
-    |> Stream.transform([], fn r, acc ->
-      if acc == [] do
-        {%{}, r |> Enum.map(fn h -> h |> String.replace_prefix("\uFEFF", "") end)}
-      else
-        {[acc |> Enum.zip(r) |> Enum.into(%{})], acc}
-      end
-    end)
-  end
-
   defp handle_stop(record) do
-    latitude = fetch_position(record, "stop_lat")
-    longitude = fetch_position(record, "stop_lon")
+    latitude = Utils.fetch_position(record, "stop_lat")
+    longitude = Utils.fetch_position(record, "stop_lon")
 
     if latitude != nil && longitude != nil do
       [
@@ -64,39 +45,11 @@ defmodule Transport.Registry.GTFS do
           latitude: latitude,
           longitude: longitude,
           projection: :utm_wgs84,
-          stop_type: record |> csv_get_with_default!("location_type", "0") |> to_stop_type()
+          stop_type: record |> Utils.csv_get_with_default("location_type", "0") |> to_stop_type()
         }
       ]
     else
       []
-    end
-  end
-
-  defp fetch_position(record, field) do
-    Map.fetch!(record, field) |> convert_text_to_float()
-  end
-
-  @doc """
-   Convert textual values to float.
-
-   iex> convert_text_to_float("")
-   nil
-   iex> convert_text_to_float("0")
-   0.0
-   iex> convert_text_to_float("0.0")
-   0.0
-   iex> convert_text_to_float("12.7")
-   12.7
-   iex> convert_text_to_float("-12.7")
-   -12.7
-   iex> convert_text_to_float("   -48.7    ")
-   -48.7
-  """
-  def convert_text_to_float(input) do
-    if input |> String.trim() != "" do
-      input |> String.trim() |> Decimal.new() |> Decimal.to_float()
-    else
-      nil
     end
   end
 
@@ -127,15 +80,5 @@ defmodule Transport.Registry.GTFS do
 
   defp entry_of_name?(name, %Unzip.Entry{file_name: file_name}) do
     file_name == name
-  end
-
-  defp csv_get_with_default!(map, field, default_value) do
-    value = Map.get(map, field)
-
-    case value do
-      nil -> default_value
-      "" -> default_value
-      v -> v
-    end
   end
 end
