+++ apps/transport/test/gtfs/utils_test.exs
@@ -1,0 +1,4 @@
+defmodule Transport.GTFS.UtilsTest do
+  use ExUnit.Case, async: true
+  doctest Transport.GTFS.Utils, import: true
+end
